Solidify utils package
